+++ src/app/app.module.ts
@@ -6,13 +6,23 @@ import { RouterModule } from '@angular/router';
 import { removeNgStyles, createNewHosts, createInputTransfer } from '@angularclass/hmr';
 
 //导入拦截器服务
-import {InterceptorService} from "ng2-interceptors/index";
-import {HttpInterceptorBackend} from './interceptor/http-interceptor-backend'
-import {HttpInterceptor} from './interceptor/http-interceptor';
-import {httpFactory} from './interceptor/http-factory';
+import { provideInterceptorService } from 'ng2-interceptors';
+import {ServerURLInterceptor} from "./ServerURLInterceptor";
+//import {InterceptorService} from "ng2-interceptors/index";
+//import {HttpInterceptorBackend} from './interceptor/http-interceptor-backend'
+//import {HttpInterceptor} from './interceptor/http-interceptor';
+//import {httpFactory} from './interceptor/http-factory';
+
+//export function interceptorFactory(xhrBackend: XHRBackend, requestOptions: RequestOptions){
+//  let service = new InterceptorService(xhrBackend, requestOptions);
+//  // Add interceptors here with service.addInterceptor(interceptor)
+//  service.addInterceptor(ServerURLInterceptor);
+//  return service;
+//}
 
 //导入核心服务
-import {CoreModule} from './core/core.module';
+import {CoreModule} from './services/coreService/core.module';
+
 
 /*
  * Platform and Environment providers/directives/pipes
@@ -25,7 +35,10 @@ import { App } from './app.component';
 import { AppState, InternalStateType } from './app.service';
 import { GlobalState } from './global.state';
 import { NgaModule } from './theme/nga.module';
+import { OaModule } from './theme/oa-them/oa.module';
 import { PagesModule } from './pages/pages.module';
+import {MessageService} from "./services/coreService/messageComponent.service";
+
 
 // Application wide providers
 const APP_PROVIDERS = [
@@ -33,7 +46,7 @@ const APP_PROVIDERS = [
   GlobalState
 ];
 
-type StoreType = {
+export type StoreType = {
   state: InternalStateType,
   restoreInputValues: () => void,
   disposeOldHosts: () => void
@@ -61,16 +74,24 @@ type StoreType = {
   providers: [ // expose our Services and Providers into Angular's dependency injection
     ENV_PROVIDERS,
     APP_PROVIDERS,
-    HttpInterceptorBackend,HttpInterceptor,
-    {
-      provide: InterceptorService,
-      useFactory: httpFactory,
-      deps: [HttpInterceptorBackend,RequestOptions]
-    }
+    //{
+    //  provide: InterceptorService,
+    //  useFactory: interceptorFactory,
+    //  deps: [XHRBackend, RequestOptions]
+    //}
+    provideInterceptorService([
+      new ServerURLInterceptor()
+    ])
+    //HttpInterceptorBackend,HttpInterceptor,
+    //{
+    //  provide: InterceptorService,
+    //  useFactory: httpFactory,
+    //  deps: [HttpInterceptorBackend,RequestOptions]
+    //}
   ]
 })
 
-export class AppModule{
+export class AppModule {
 
   constructor(public appRef: ApplicationRef, public appState: AppState) {
   }