+++ src/app/pages/sale/components/info/provision/provision.component.ts
@@ -1,9 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'provision',
-  template: require('./provision.html'),
-  styles: []
-})
-
-export class ProvisionComponent{}