+++ src/app/theme/components/baContentTop/baContentTop.component.ts
@@ -4,8 +4,8 @@ import {GlobalState} from '../../../global.state';
 
 @Component({
   selector: 'ba-content-top',
-  styles: [require('./baContentTop.scss')],
-  template: require('./baContentTop.html'),
+  styleUrls: ['./baContentTop.scss'],
+  templateUrl: './baContentTop.html',
 })
 export class BaContentTop {
 
@@ -14,7 +14,6 @@ export class BaContentTop {
   constructor(private _state:GlobalState) {
     this._state.subscribe('menu.activeLink', (activeLink) => {
       if (activeLink) {
-        console.log(activeLink);
         this.activePageTitle = activeLink.title;
       }
     });