+++ src/app/theme/nga.module.ts
@@ -2,9 +2,7 @@ import { NgModule, ModuleWithProviders }      from '@angular/core';
 import { CommonModule }  from '@angular/common';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
-import {AgGridModule} from 'ag-grid-ng2/main';
-import {TreeModule} from "angular2-tree-component/dist/angular2-tree-component";
-import { DropdownModule, ModalModule } from 'ng2-bootstrap/ng2-bootstrap';
+import { NgUploaderModule } from 'ngx-uploader';
 
 import {
   BaThemeConfig
@@ -28,20 +26,7 @@ import {
   BaMultiCheckbox,
   BaPageTop,
   BaPictureUploader,
-  BaSidebar,
-  Ng2Tab,
-  ActionBar,
-  TextAlertComponent,
-  PagesComponent,
-  PagesButtonComponent,
-  CalenderComponent,
-  MessageComponent,
-  ScheduleComponent,
-  NgSelectComponent,
-  CustomerSelectComponent,
-  ProductSelectComponent,
-  SearchComponent,
-  CostComponent
+  BaSidebar
 } from './components';
 
 import { BaCardBlur } from './components/baCard/baCardBlur.directive';
@@ -55,32 +40,20 @@ import {
 import {
   BaAppPicturePipe,
   BaKameleonPicturePipe,
-  BaProfilePicturePipe,
-  CountryPipe,
-  ProjectPipe,
-  UsersPipe,
-  PaymentPipe,
-  CurrencyPipe,
-  OrderStatusPipe,
-  TransportPipe
+  BaProfilePicturePipe
 } from './pipes';
 
 import {
   BaImageLoaderService,
+  BaMenuService,
   BaThemePreloader,
-  BaThemeSpinner,
-  TextAlertService
+  BaThemeSpinner
 } from './services';
 
 import {
   EmailValidator,
   EqualPasswordsValidator
 } from './validators';
-
-import {
-  AuthGuard
-} from './guards'
-
 
 const NGA_COMPONENTS = [
   BaAmChart,
@@ -96,20 +69,7 @@ const NGA_COMPONENTS = [
   BaMultiCheckbox,
   BaPageTop,
   BaPictureUploader,
-  BaSidebar,
-  Ng2Tab,
-  ActionBar,
-  TextAlertComponent,
-  PagesComponent,
-  PagesButtonComponent,
-  CalenderComponent,
-  MessageComponent,
-  ScheduleComponent,
-  NgSelectComponent,
-  CustomerSelectComponent,
-  ProductSelectComponent,
-  SearchComponent,
-  CostComponent
+  BaSidebar
 ];
 
 const NGA_DIRECTIVES = [
@@ -122,21 +82,14 @@ const NGA_DIRECTIVES = [
 const NGA_PIPES = [
   BaAppPicturePipe,
   BaKameleonPicturePipe,
-  BaProfilePicturePipe,
-  CountryPipe,
-  ProjectPipe,
-  UsersPipe,
-  PaymentPipe,
-  CurrencyPipe,
-  OrderStatusPipe,
-  TransportPipe
+  BaProfilePicturePipe
 ];
 
 const NGA_SERVICES = [
   BaImageLoaderService,
   BaThemePreloader,
   BaThemeSpinner,
-  TextAlertService
+  BaMenuService
 ];
 
 const NGA_VALIDATORS = [
@@ -144,29 +97,23 @@ const NGA_VALIDATORS = [
   EqualPasswordsValidator
 ];
 
-const NGA_GUARDS = [
-  AuthGuard
-]
-
 @NgModule({
   declarations: [
     ...NGA_PIPES,
     ...NGA_DIRECTIVES,
-    ...NGA_COMPONENTS,
+    ...NGA_COMPONENTS
   ],
   imports: [
     CommonModule,
     RouterModule,
     FormsModule,
-    AgGridModule,
-    TreeModule,
     ReactiveFormsModule,
-    DropdownModule, ModalModule
+    NgUploaderModule,
   ],
   exports: [
     ...NGA_PIPES,
     ...NGA_DIRECTIVES,
-    ...NGA_COMPONENTS,
+    ...NGA_COMPONENTS
   ]
 })
 export class NgaModule {