+++ src/app/theme/oa-them/components/scheduleComponent/schedule.component.ts
@@ -1,0 +1,13 @@
+import {Component,Input} from '@angular/core';
+
+@Component({
+  selector: 'ng-schdule',
+  templateUrl: './schedule.component.html',
+  styleUrls: ['./schedule.component.scss']
+})
+
+export class ScheduleComponent{
+  @Input() schduleData: {};
+
+
+}